Separate munge for PBS and Slurm tests
